+++ src/repositories/Maps.ts
@@ -9,7 +9,7 @@ import { KeyStore } from "./KeyStore";
 // Stores
 export const hpkeKeysMap = new KeyStore("HpkeKeys");
 export const hpkeKeyIdMap = new LastestItemStore<number, string>("HpkeKeyids");
-export const keyReleaseMapName = "public:ccf.gov.policies.key_release";
+export const keyReleaseMapName = "public:policies.key_release";
 export const keyReleasePolicyMap = ccf.kv[keyReleaseMapName];
 export const settingsMapName = "public:policies.settings";
 export const settingsPolicyMap = ccf.kv[settingsMapName];
